requirements.txt
flask>=3.0
flask-cors>=4.0
gunicorn>=21.2
torch>=2.0
timm>=0.9
albumentations>=1.3
opencv-python>=4.8
numpy>=1.24
xgboost>=1.7
Pillow>=10.0
